+++ CHANGELOG.md
@@ -14,6 +14,14 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.1.1.26] - 2021-01-26
+
+### Changed
+- submenu is displayed on hover
+
+### Fixed
+- proper calculation production for today if last state is production
+
 ## [2021.1.1.25] - 2021-01-25
 
 ### Added